fix(akamai): only attempt to apply revisions that affect resources
If the resource folder contains extra files that do not directly concern
the execution of bossman, changing them (adding, moving etc...) would
cause the revision to be added to the apply list, but the property
resource would raise an assertionerror because it should not happen.
This fix prevents all known cases for that assertion error.