fix(git): reduce chattiness when a commit id cannot be resolved
Before this change, bossman would log with a high priority whenever
a commit referenced in an asset (e.g. property version notes) did
not exist. This situation is fairly common when history is rewritten
which is legitimate for short-lived branches that need to be rebased.

Reducing the priority of the message will reduce confusion for users
and makes sense since the error is correctly handled by the software
and does not require operator involvement.